refactor: extract settings metadata contracts
